feat(dashboard): rutas directas /resumen, /balance y /analisis para las pestañas del dashboard
- El contenido de la home se mueve a components/dashboard/dashboard-home.tsx
- /resumen, /balance y /analisis abren el dashboard con su pestaña activa
- Al cambiar de pestaña la URL se sincroniza sin recargar la página
