+++ app/balance/page.tsx
@@ -1,0 +1,7 @@
+"use client"
+
+import { DashboardHome } from "@/components/dashboard/dashboard-home"
+
+export default function BalancePage() {
+  return <DashboardHome initialTab="actividad" />
+}
